requirements.txt
icalendar==5.0.4
py-cord==2.2.2
python-dotenv==1.0.0
pymongo==4.3.3
